add ``delayed down'' functionality
This commit makes it the case that, if a janus key is held down for
more than the max delay specified in the config file, then its 2nd
function is sent down.

The main reason why I wanted to implement this feature was that I
couldn't zoom in the browser using left ctrl + mouse scroll.